fix(#22): create unified PriceInput component and convert prices to integer storage
This commit implements a comprehensive solution for handling prices across the application:

## Changes

### New Components
- **PriceInput component**: A reusable input component for monetary values with:
  - Always displays 2 decimals (adds trailing zeros as needed)
  - Currency symbol (€) displayed in input
  - Thousand separators based on locale (fr-FR)
  - `inputmode="decimal"` for mobile keyboards
  - Automatic formatting on blur
  - Text selection on focus for easy editing

### Price Storage Model
- **Changed from float to integer (cents)**:
  - Prevents floating-point arithmetic errors
  - Display: "€123,45" (formatted, user-friendly)
  - Store: 12345 (cents, integer)

### Database Migration
- Created migration to convert all price fields from numeric(10,2) to bigint
- Products.price: now stores cents
- Invoice items (unit_price, line_total): now store cents
- Invoices.total_amount: now stores cents

### Utility Functions
- `centsToCurrency()`: Convert cents to currency units
- `currencyToCents()`: Convert currency units to cents
- `parsePrice()`: Parse user input strings to cents
- `centsToCurrencyString()`: Format cents as currency string

### Updated Components & Pages
- Product forms (new & edit): Use PriceInput component
- Invoice ArticlesBlock: Use PriceInput for line item prices
- Invoice creation page: Handle cents conversion in calculations
- All display pages (products, invoices): Display cents as formatted currency
- PDF generation: Convert cents to currency for display

### Type & Validation Updates
- Updated TypeScript types to reflect integer storage
- Updated Zod schemas to validate integer prices
- Added comments documenting units (cents vs currency)

All changes maintain backward compatibility with existing data through the database migration.

+++ src/lib/validation.ts
@@ -4,7 +4,7 @@ export const productSchema = z.object({
   id: z.string().uuid().optional(),
   name: z.string().min(1),
   description: z.string().optional().nullable(),
-  price: z.number().nonnegative(),
+  price: z.number().int().nonnegative(), // price in cents (integer)
   image_url: z.string().url().or(z.literal("")).optional().nullable(),
 });
 
@@ -19,16 +19,16 @@ export const clientSchema = z.object({
 export const invoiceItemSchema = z.object({
   product_id: z.string().uuid(),
   name: z.string().min(1),
-  quantity: z.number().int().positive(),
-  price: z.number().nonnegative(),
-  total: z.number().nonnegative(),
+  quantity: z.number().positive(), // can be fractional (e.g., 1.5 hours)
+  price: z.number().int().nonnegative(), // unit price in cents (integer)
+  total: z.number().int().nonnegative(), // line total in cents (integer)
 });
 
 export const invoiceSchema = z.object({
   id: z.string().uuid().optional(),
   client_id: z.string().uuid(),
   items: z.array(invoiceItemSchema),
-  total_amount: z.number().nonnegative(),
+  total_amount: z.number().int().nonnegative(), // total amount in cents (integer)
   status: z.enum(["draft", "sent", "paid", "overdue"]).default("draft"),
   issue_date: z.string(),
   due_date: z.string().optional(),
